move ApiAttributes to Api::Serializable
